Inline authAddPerson & authUpgradePerson into people/router
